extract python module evaluator runner
